initial commit
The first run through showing how Entity Framework 6 and FluentMigrator
can work together in tandum. This is more suitable for working in a
team, where the EF migrations break down due to the binary hashing.

The downside is you have to handcraft your migrations, but not a very
significant one in my opinion.